Implementation complete.
There where synchronization problems. __syncthreads() only syncs thread in
the current block. The kernel is now split in two and a call to
cudaDeviceSynchronize() is doen to sync the whole device and to make sure
that all squares are created before the dimaonds are created. The funny
pattern and the artifacts are therefore removed.

There still work to do regarding optimization. The random init is now done
at application start and not for every frame. Gigantic performance boost
compared to earlier impl.